OTP verification by Laxman

+++ src/main/java/com/credit/entity/CreditCard.java
@@ -6,8 +6,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -19,7 +17,7 @@ public class CreditCard {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Integer cardId;
-	private Integer cardNumber;
+	private Long cardNumber;
 	private Integer cvv;
 	private LocalDate expireDate;
 	private Integer availableLimit ;
